LayoutTests/fast/dom/constructors-cached-navigate-expected.txt


PASS successfullyParsed is true

TEST COMPLETE
PASS testFrame.contentWindow.Image is testFrame.contentWindow.Image
PASS testFrame.contentWindow.Image !== window.Image is true
PASS testFrame.contentWindow.Image.testProperty is "property set successfully"
PASS testFrame.contentWindow.MessageChannel is testFrame.contentWindow.MessageChannel
FAIL testFrame.contentWindow.MessageChannel !== window.MessageChannel should be true. Was false.
FAIL Testing MessageChannel threw TypeError: Result of expression 'testFrame.contentWindow[constructor]' [undefined] is not an object.
PASS testFrame.contentWindow.Option is testFrame.contentWindow.Option
PASS testFrame.contentWindow.Option !== window.Option is true
PASS testFrame.contentWindow.Option.testProperty is "property set successfully"
PASS testFrame.contentWindow.XMLHttpRequest is testFrame.contentWindow.XMLHttpRequest
PASS testFrame.contentWindow.XMLHttpRequest !== window.XMLHttpRequest is true
PASS testFrame.contentWindow.XMLHttpRequest.testProperty is "property set successfully"
PASS testFrame.contentWindow.Audio is testFrame.contentWindow.Audio
PASS testFrame.contentWindow.Audio !== window.Audio is true
PASS testFrame.contentWindow.Audio.testProperty is "property set successfully"
PASS testFrame.contentWindow.Image is testFrame.contentWindow.Image
PASS testFrame.contentWindow.Image !== window.Image is true
PASS testFrame.contentWindow.Image !== storedConstructors.Image is true
PASS storedConstructors.Image.testProperty is "property set successfully"
PASS testFrame.contentWindow.Image.testProperty is undefined.
PASS testFrame.contentWindow.Image.cachedOnOwnerDocument is true
PASS testFrame.contentWindow.MessageChannel is testFrame.contentWindow.MessageChannel
FAIL testFrame.contentWindow.MessageChannel !== window.MessageChannel should be true. Was false.
FAIL testFrame.contentWindow.MessageChannel !== storedConstructors.MessageChannel should be true. Was false.
FAIL storedConstructors.MessageChannel.testProperty should be property set successfully. Threw exception TypeError: Result of expression 'storedConstructors.MessageChannel' [undefined] is not an object.
FAIL testFrame.contentWindow.MessageChannel.testProperty should be undefined. Threw exception TypeError: Result of expression 'testFrame.contentWindow.MessageChannel' [undefined] is not an object.
FAIL testFrame.contentWindow.MessageChannel.cachedOnOwnerDocument should be true. Threw exception TypeError: Result of expression 'testFrame.contentWindow.MessageChannel' [undefined] is not an object.
PASS testFrame.contentWindow.Option is testFrame.contentWindow.Option
PASS testFrame.contentWindow.Option !== window.Option is true
PASS testFrame.contentWindow.Option !== storedConstructors.Option is true
PASS storedConstructors.Option.testProperty is "property set successfully"
PASS testFrame.contentWindow.Option.testProperty is undefined.
PASS testFrame.contentWindow.Option.cachedOnOwnerDocument is true
PASS testFrame.contentWindow.XMLHttpRequest is testFrame.contentWindow.XMLHttpRequest
PASS testFrame.contentWindow.XMLHttpRequest !== window.XMLHttpRequest is true
PASS testFrame.contentWindow.XMLHttpRequest !== storedConstructors.XMLHttpRequest is true
PASS storedConstructors.XMLHttpRequest.testProperty is "property set successfully"
PASS testFrame.contentWindow.XMLHttpRequest.testProperty is undefined.
PASS testFrame.contentWindow.XMLHttpRequest.cachedOnOwnerDocument is true
PASS testFrame.contentWindow.Audio is testFrame.contentWindow.Audio
PASS testFrame.contentWindow.Audio !== window.Audio is true
PASS testFrame.contentWindow.Audio !== storedConstructors.Audio is true
PASS storedConstructors.Audio.testProperty is "property set successfully"
PASS testFrame.contentWindow.Audio.testProperty is undefined.
PASS testFrame.contentWindow.Audio.cachedOnOwnerDocument is true
